+++ src/index.js
@@ -5,4 +5,9 @@ import Button from './button';
 export * from './icon-button';
 export * from './spinners';
 export * from './slider';
+export * from './tile-list';
+
+export * from './utils/hooks/use-dialog';
+export * from './dialog';
+
 export { Input, Button, Icon };