+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 1.3.0 [unreleased]
+
+-   **Changed** dialog API to make primary button optional [#53](https://github.com/egfanboy/t-components/pull/53)
+
+## 🦄 New Features 🦄
+
+-   **Created** dialog component [#52](https://github.com/egfanboy/t-components/pull/52)
+-   **Created** tile list component [#49](https://github.com/egfanboy/t-components/pull/49)
+
 # 1.2.0 [08/16/2020]
 
 ## ✨Enhancement ✨